Add foundation for backend integration tests (API and otherwise)
The utility code makes it possible to start an isolated Tobira with
separate DB, separate Meili index, a specific config and listening on
a Unix socket. With this, it's fairly easy to send HTTP requests to
the running Tobira and make sure the response is as expected.

There is still lots missing and more utility functions will be added
once we write more tests.